Updating changes

+++ IDEtalk/src/idea/jetbrains/communicator/idea/toolWindow/UserListComponentImpl.java
@@ -364,7 +364,7 @@ public class UserListComponentImpl implements UserListComponent, Disposable {
 
     private boolean problem(final String resourceCode) {
       EdtExecutorService.getScheduledExecutorInstance().schedule(() -> Callout.showText(myTree, myTree.getEditingPath(), Callout.SOUTH_WEST,
-                                                                                        CommunicatorStrings.getMsg(resourceCode)), (long)100, TimeUnit.MILLISECONDS);
+                                                                                        CommunicatorStrings.getMsg(resourceCode)), 100, TimeUnit.MILLISECONDS);
       return false;
     }
